core/src/main/java/nl/matsgemmeke/battlegrounds/item/equipment/mechanism/EquipmentMechanismFactory.java
package nl.matsgemmeke.battlegrounds.item.equipment.mechanism;

import dev.dejvokep.boostedyaml.block.implementation.Section;
import nl.matsgemmeke.battlegrounds.game.GameContext;
import nl.matsgemmeke.battlegrounds.item.InvalidItemConfigurationException;
import org.jetbrains.annotations.NotNull;

public class EquipmentMechanismFactory {

    public EquipmentMechanism make(@NotNull Section section, @NotNull GameContext context) {
        String type = section.getString("type");

        if (type == null) {
            throw new InvalidItemConfigurationException("Equipment mechanism type must be defined!");
        }

        EquipmentMechanismType equipmentMechanismType;

        try {
            equipmentMechanismType = EquipmentMechanismType.valueOf(type);
        } catch (IllegalArgumentException e) {
            throw new InvalidItemConfigurationException("Equipment mechanism type \"" + type + "\" is invalid!");
        }

        switch (equipmentMechanismType) {
            case EXPLOSION -> {
                return new ExplosionMechanism(context);
            }
        }

        throw new InvalidItemConfigurationException("Unknown equipment mechanism type \"" + type + "\"!");
    }
}
